Separate methods in a new class to create and display books
